requirements.txt
pytest
torch
torchvision
numpy
pillow
scipy

viam-sdk
altgraph==0.17.4
astroid==3.3.8
coloredlogs==15.0.1
dill==0.3.9
dnspython==2.7.0
exceptiongroup==1.2.2
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2024.12.0
googleapis-common-protos==1.66.0
grpclib==0.4.7
h2==4.1.0
hpack==4.1.0
humanfriendly==10.0
hyperframe==6.1.0
iniconfig==2.0.0
isort==6.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==5.29.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
pylint==3.3.4
pymongo==4.11
pytest==8.3.4
pytest-asyncio==0.25.3
scipy==1.15.1
sympy==1.13.1
tabulate==0.9.0
tomli==2.2.1
tomlkit==0.13.2
types-tabulate==0.9.0.20241207
typing_extensions==4.12.2
viam-sdk==0.38.0
